+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2024-05-13T07:49:05Z",
+  "generated_at": "2024-05-15T13:37:47Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -77,10 +77,25 @@
     }
   ],
   "results": {
-    "examples/kubernetes/cos-s3-csi-pvc-secret.yaml": [
+    "examples/kubernetes/cos-s3-csi-rclone-secret.yaml": [
       {
         "hashed_secret": "630f59d934b0bc814e3928cf48be707f28e4718d",
-        "is_secret": false,
+        "is_verified": false,
+        "line_number": 13,
+        "type": "Base64 High Entropy String",
+        "verified_result": null
+      },
+      {
+        "hashed_secret": "d08f88df745fa7950b104e4a707a31cfce7b5841",
+        "is_verified": false,
+        "line_number": 19,
+        "type": "Secret Keyword",
+        "verified_result": null
+      }
+    ],
+    "examples/kubernetes/cos-s3-csi-s3fs-secret.yaml": [
+      {
+        "hashed_secret": "630f59d934b0bc814e3928cf48be707f28e4718d",
         "is_verified": false,
         "line_number": 13,
         "type": "Base64 High Entropy String",
@@ -96,26 +111,10 @@
     ],
     "pkg/driver/controllerserver.go": [
       {
-        "hashed_secret": "39f69c278f46165447f30d10acf54277aaa3d5fc",
-        "is_secret": false,
-        "is_verified": false,
-        "line_number": 60,
-        "type": "Secret Keyword",
-        "verified_result": null
-      },
-      {
-        "hashed_secret": "f75b33f87ffeacb3a4f793a09693e672e07449ff",
-        "is_secret": false,
-        "is_verified": false,
-        "line_number": 86,
-        "type": "Secret Keyword",
-        "verified_result": null
-      },
-      {
         "hashed_secret": "c7c6508b19455e3e8040e60e9833fbede92e5d8e",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 212,
+        "line_number": 155,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -134,7 +133,7 @@
         "hashed_secret": "39f69c278f46165447f30d10acf54277aaa3d5fc",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 123,
+        "line_number": 89,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -142,7 +141,7 @@
         "hashed_secret": "2ace62c1befa19e3ea37dd52be9f6d508c5163e6",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 242,
+        "line_number": 244,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -152,7 +151,7 @@
         "hashed_secret": "39f69c278f46165447f30d10acf54277aaa3d5fc",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 77,
+        "line_number": 78,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -191,18 +190,18 @@
         "verified_result": null
       }
     ],
-    "pkg/utils/utils.go": [
+    "pkg/utils/driver_utils.go": [
       {
         "hashed_secret": "c7c6508b19455e3e8040e60e9833fbede92e5d8e",
         "is_verified": false,
-        "line_number": 236,
+        "line_number": 124,
         "type": "Secret Keyword",
         "verified_result": null
       },
       {
         "hashed_secret": "788f47a335cc4d15e0cdbe8850580c6468303558",
         "is_verified": false,
-        "line_number": 283,
+        "line_number": 171,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -211,7 +210,7 @@
       {
         "hashed_secret": "d5cd5ec53cf11ad72f24451e6c077d94d0481046",
         "is_verified": false,
-        "line_number": 61,
+        "line_number": 63,
         "type": "Secret Keyword",
         "verified_result": null
       }
